gp-import-bends: refactored, moved old implementation to separate files
