+++ src/importexport/guitarpro/internal/guitarbendimport/splitchord/bendchorddurationsplitter.cpp
