+++ src/importexport/guitarpro/internal/guitarbendimport/splitchord/bendchorddurationsplitter.h
