+++ src/importexport/guitarpro/internal/guitarbendimport/splitchord/benddatacontextsplitchord.h
@@ -5,7 +5,7 @@
  * MuseScore Studio
  * Music Composition & Notation
  *
- * Copyright (C) 2024 MuseScore Limited
+ * Copyright (C) 2025 MuseScore Limited
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -22,23 +22,14 @@
 
 #pragma once
 
-#include "benddatacontext.h"
-#include "splitchord/benddataprocessorsplitchord.h"
-
-namespace mu::engraving {
-class Note;
-class Score;
-}
+#include <engraving/types/types.h>
+#include "../guitarbendimporttypes.h"
 
 namespace mu::iex::guitarpro {
-class BendDataProcessor
-{
-public:
-    BendDataProcessor(mu::engraving::Score* score);
-    void processBends(const BendDataContext& bendDataCtx);
-
-private:
-    mu::engraving::Score* m_score = nullptr;
-    std::unique_ptr<BendDataProcessorSplitChord> m_bendDataProcessorSplitChord;
+struct BendDataContextSplitChord {
+    std::unordered_map<mu::engraving::track_idx_t, std::map<int, BendChordData> > bendDataByEndTick;
+    std::unordered_map<mu::engraving::track_idx_t, std::map<int, std::vector<mu::engraving::Fraction> > > bendChordDurations;
+    std::unordered_map<mu::engraving::track_idx_t, std::set<mu::engraving::Fraction> > reduntantChordTicks;
+    std::unordered_map<mu::engraving::track_idx_t, std::set<mu::engraving::Fraction> > chordTicksForTieBack;
 };
 } // mu::iex::guitarpro
